refactor db.close() out of inner calls
Whenever managing a resource that requires an open/close,
such as a database, follow the [RAII] idiom. Loosly speaking,
close the resource in the same scope as it was opened.  Here
I use the try/finally pattern to ensure that db.close() is called
even if an exception is thrown during execution of 'start(...)'.

Later I'll refactor the DB connection code out into its own object
so that we can do something like

with open_connect(...) as cursor:
	# use cursor